removed parameter input via command line arguments; reworked histogram creation

+++ src/main/java/morris_water_maze/parameter/MouseParameterAccessor.java
@@ -2,11 +2,16 @@ package morris_water_maze.parameter;
 
 import morris_water_maze.model.StartingSide;
 
+
 public interface MouseParameterAccessor
 {
-    int getMaximumMouseSwimmingDuration();
+    double getMaximumMouseSwimmingDuration();
+    
     double getMouseTrainingLevel();
+    
     double getStepLengthBias();
+    
     StartingSide getStartingSide();
+    
     double mouseSpeed();
 }
